+++ src/test/kotlin/icu/windea/pls/lang/intentions/script/PutMembersIntentionsTest.kt
@@ -1,0 +1,24 @@
+package icu.windea.pls.lang.intentions.script
+
+import com.intellij.testFramework.TestDataPath
+import com.intellij.testFramework.fixtures.BasePlatformTestCase
+import icu.windea.pls.test.clearIntegrationTest
+import icu.windea.pls.test.markIntegrationTest
+import org.junit.After
+import org.junit.Before
+import org.junit.runner.RunWith
+import org.junit.runners.JUnit4
+
+@RunWith(JUnit4::class)
+@TestDataPath("\$CONTENT_ROOT/testData")
+class PutMembersIntentionsTest: BasePlatformTestCase() {
+    override fun getTestDataPath() = "src/test/testData"
+
+    @Before
+    fun setup() = markIntegrationTest()
+
+    @After
+    fun clear() = clearIntegrationTest()
+
+    // TODO ...
+}